fix errs

+++ source/library/actl/project/Project.py
@@ -31,11 +31,14 @@ def recursice_update(base, update_dict):
 
 
 class Project:
+	this = None
 	__data = collections.OrderedDict()
 	__data['handlers'] = {}
 	add_syntax = property(lambda self: self[('rules',)].add)
 
 	def __init__(self, projectf=None, source=None, data=None):
+		if type(self).this is None:
+			type(self).this = self
 		self.__data = copy.deepcopy(self.__data)
 		self.source = []
 		if data is not None:
@@ -69,21 +72,18 @@ class Project:
 	def __iter__(self):
 		return iter(self.__data.items())
 
-	def __contains__(self, key):
-		return key in self.__data
-
-	def __getitem__(self, keys):
-		value = self.__data
+	def __getitem__(self, keys, data=None):
+		value = data or self.__data
 		for key in keys:
 			value = value[key]
-		if isinstance(value, VirtualProjectItem):
+		if isinstance(value, VirtualProjectItem) and hasattr(value, 'get'):
 			value = value.get()
 		return value
 
-	def __setitem__(self, keys, value):
-		base = self[keys[:-1]]
+	def __setitem__(self, keys, value, base=None):
+		base = base or self[keys[:-1]]
 		key = keys[-1]
-		if (key in base) and isinstance(base[key], VirtualProjectItem):
+		if (key in base) and isinstance(base[key], VirtualProjectItem) and hasattr(value, 'set'):
 			base[key].set(value)
 		else:
 			base[key] = value
@@ -91,7 +91,7 @@ class Project:
 	def __delitem__(self, keys):
 		base = self[keys[:-1]]
 		key = keys[-1]
-		if isinstance(base[key], VirtualProjectItem):
+		if isinstance(base[key], VirtualProjectItem) and hasattr(base[key], 'delete'):
 			base[key].delete()
 		else:
 			del base[key]
@@ -103,16 +103,42 @@ class Project:
 			return value
 		return decorator
 
+	def __repr__(self):
+		return f'{type(self).__name__}({self.__data})'
+
+
+class SubProject(Project):
+	def __init__(self, project, source):
+		self.__data = {}
+		super().__init__(data=project, source=source)
+
+	def update(self, data):
+		self.__data.update(data)
+
+	def __iter__(self):
+		return iter(self.__data.items())
+
+	def __getitem__(self, keys):
+		try:
+			return super().__getitem__(keys)
+		except KeyError:
+			return super().__getitem__(keys, data=self.__data)
+
+	def __setitem__(self, keys, value):
+		super().__setitem__(keys, value, base=self[keys[:-1]])
+
+	def __repr__(self):
+		return f'{type(self).__name__}(project={super().__repr__()}, data={self.__data})'
 
 class VirtualProjectItem:
-	def get(self):
-		raise NotImplementedError
+	def __init__(self, project, key, value):
+		self._project = project
+		self._key = key
+		self._value = value
+		self._project[(self._key,)] = self
 
-	def set(self, value):
-		raise NotImplementedError
-
-	def delete(self, value):
-		raise NotImplementedError
+	def __repr__(self):
+		return f'{type(self).__name__}(key={self._key}, value={self._value})'
 
 
 @Project.add_handler('import')
@@ -130,13 +156,29 @@ def _(project, key, body):
 	body = f'def {key}:\n  {body}'
 	lc_scope = {}
 	try:
-		exec(body, {'this': project}, lc_scope)
+		exec(body, {'this': type(project).this}, lc_scope)
 	except:
 		print(f"{project}['{key}']")
 		print(body)
 		raise
 	key = key[:key.find('(')]
 	project[(key,)] = lc_scope[key]
+
+
+@Project.add_handler('pypropl')
+class ProjectItemPyPropLazy(VirtualProjectItem):
+	def __init__(self, project, key, value):
+		super().__init__(project, key, value)
+
+	def get(self):
+		try:
+			return self.__self
+		except AttributeError:
+			subproject = SubProject(
+				self._project, [{key: value} for key, value in self._value.items()]
+			)
+			self.__self = subproject[('get',)]()
+			return self.get()
 
 
 class LinkerLayer:
